✨feat: 포스트 저장 시, 스티커 리스트형태로 저장

+++ src/main/java/com/justdo/glue/sticker/domain/poststicker/PostSticker.java
@@ -27,12 +27,12 @@ public class PostSticker extends BaseTimeEntity {
     @Column(nullable = false, columnDefinition = "INT")
     private int yLocation;
 
-    @Column(nullable = false, columnDefinition = "INT")
-    private int width;
+    @Column(nullable = false, columnDefinition = "DOUBLE")
+    private double width;
 
-    @Column(nullable = false, columnDefinition = "INT")
-    private int height;
+    @Column(nullable = false, columnDefinition = "DOUBLE")
+    private double height;
 
-    @Column(nullable = false, columnDefinition = "INT")
-    private int angle;
+    @Column(nullable = false, columnDefinition = "DOUBLE")
+    private double angle;
 }
